Fix the last merge and make the t.bkd-expired work correctly in the 3.3.x tree.  When a bkd has an expired license (or no license at all) most all bk operations should continue to work correctly because bk clients send the current license across the wire to the bkd.
This wasn't working correctly because BK_AUTH_HMAC didn't get validated
in bkd.c.  The hashed region needs to end before the command gets
executed for the code in bkd.c to work.   That means we need to use
the 'extra' field on send_file().

bk: 4458f980Mbra1GF4RiAR-Sl6XBXTJQ

+++ src/push.c
@@ -159,9 +159,10 @@ err:		freeLines(envVar, free);
 private int
 send_part1_msg(remote *r, char **envVar)
 {
-	char	*cmd, buf[MAXPATH];
 	FILE 	*f;
-	int	gzip, rc;
+	int	gzip, rc, i;
+	char	*probef;
+	char	buf[MAXPATH];
 
 	/*
 	 * If we are using ssh/rsh do not do gzip ourself
@@ -181,12 +182,18 @@ send_part1_msg(remote *r, char **envVar)
 	fputs("\n", f);
 	fclose(f);
 
-	cmd = aprintf("bk _probekey  >> %s", buf);
-	system(cmd);
-	free(cmd);
-
-	rc = send_file(r, buf, 0);
-	unlink(buf);
+	probef = bktmp(0, 0);
+	unless (rc = sysio(0, probef, 0, "bk", "_probekey", SYS)) {
+		rc = send_file(r, buf, size(probef));
+		unlink(buf);
+		f = fopen(probef, "rb");
+		while ((i = fread(buf, 1, sizeof(buf), f)) > 0) {
+			writen(r->wfd, buf, i);
+		}
+		fclose(f);
+	}
+	unlink(probef);
+	free(probef);
 	return (rc);
 }
 
@@ -410,11 +417,10 @@ send_end_msg(remote *r, char *msg, char *rev_list, char **envVar)
 	if (gzip) fprintf(f, " -z%d", opts.gzip);
 	unless (opts.doit) fprintf(f, " -n");
 	fputs("\n", f);
-
-	fputs(msg, f);
 	fclose(f);
 
-	rc = send_file(r, msgfile, 0);
+	rc = send_file(r, msgfile, strlen(msg));
+	writen(r->wfd, msg, strlen(msg));
 	unlink(msgfile);
 	unlink(rev_list);
 	return (0);
@@ -464,22 +470,23 @@ send_patch_msg(remote *r, char rev_list[], char **envVar)
 	}
 	unless (opts.doit) fprintf(f, " -n");
 	fputs("\n", f);
-	fprintf(f, "@PATCH@\n");
 	fclose(f);
 
 	/*
 	 * Httpd wants the message length in the header
 	 * We have to compute the patch size before we sent
+	 * 8 is the size of "@PATCH@"
 	 * 6 is the size of "@END@" string
 	 */
 	if (r->type == ADDR_HTTP) {
 		m = patch_size(gzip, rev_list);
 		assert(m > 0);
-		extra = m + 6;
+		extra = m + 8 + 6;
 	}
 
 	rc = send_file(r, msgfile, extra);
 
+	writen(r->wfd, "@PATCH@\n", 8);
 	n = genpatch(gzip, r->wfd, rev_list);
 	if ((r->type == ADDR_HTTP) && (m != n)) {
 		fprintf(opts.out,
